% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprohealth-package.R
\docType{package}
\name{reprohealth-package}
\alias{reprohealth}
\alias{reprohealth-package}
\title{reprohealth: Data on Global Women's Health from World Bank and Gapminder}
\description{
An excerpt of the data available at Gapminder.org and data.worldbank.org. For each of 142
countries, the package provides values for life expectancy, contraception,
fertility rates, maternal mortality, and school enrollment, every five years, from 1952 to 2007.
}
\author{
\strong{Maintainer}: Alison Hill \email{apresstats@gmail.com}

}
\keyword{internal}
