% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprohealth-package.R
\docType{package}
\name{reprohealth}
\alias{reprohealth}
\alias{reprohealth-package}
\title{An R data package that includes reproductive health indicators for global monitoring from the World Bank and Gapminder.
Contribute to this package at \url{https://github.com/apreshill/reprohealth}.}
\description{
An R data package that includes reproductive health indicators for global monitoring from the World Bank and Gapminder.
Contribute to this package at \url{https://github.com/apreshill/reprohealth}.
}
\examples{
# Example usage:
library(reprohealth)
head(fertility)

# All information about any data set can be found in the help file:
?fertility

# To view a list of all data sets:
data(package = "reprohealth")

# To browse all vignettes:
browseVignettes(package = "reprohealth")
}
